requirements.txt
streamlit
langchain_community
langchain-huggingface
langchain-groq
pypdf
faiss-cpu
python-dotenv